# Agentic_outputs/5jun/phase_3_domain_design/api_contracts.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCMS API",
    "version": "1.0.0",
    "description": "Generic Docker Container Management System REST API. Provides full lifecycle management of containers, images, networks, volumes, clusters, monitoring, logging, and user administration across multi-host Docker Swarm deployments."
  },
  "servers": [
    {
      "url": "https://api.dcms.internal/v1",
      "description": "Production API Gateway"
    },
    {
      "url": "https://api.staging.dcms.internal/v1",
      "description": "Staging API Gateway"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/containers": {
      "get": {
        "summary": "List containers with optional filters and pagination",
        "tags": ["containers"],
        "operationId": "listContainers",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by namespace (label dcms.namespace)"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["created", "running", "paused", "restarting", "removing", "exited", "dead"]
            },
            "description": "Filter by container status"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of containers",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginatedResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Container" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create and optionally start a new container",
        "tags": ["containers"],
        "operationId": "createContainer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContainerCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Container created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Container" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/containers/{id}": {
      "get": {
        "summary": "Get detailed information about a specific container",
        "tags": ["containers"],
        "operationId": "getContainer",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" }
        ],
        "responses": {
          "200": {
            "description": "Container detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Container" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Remove a container (must be stopped first unless force=true)",
        "tags": ["containers"],
        "operationId": "deleteContainer",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" },
          {
            "name": "force",
            "in": "query",
            "schema": { "type": "boolean", "default": false },
            "description": "Force removal of a running container"
          },
          {
            "name": "remove_volumes",
            "in": "query",
            "schema": { "type": "boolean", "default": false },
            "description": "Remove anonymous volumes attached to the container"
          }
        ],
        "responses": {
          "200": {
            "description": "Container removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/containers/{id}/start": {
      "post": {
        "summary": "Start a stopped container",
        "tags": ["containers"],
        "operationId": "startContainer",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" }
        ],
        "responses": {
          "200": {
            "description": "Container started",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Container" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/containers/{id}/stop": {
      "post": {
        "summary": "Stop a running container",
        "tags": ["containers"],
        "operationId": "stopContainer",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" },
          {
            "name": "timeout",
            "in": "query",
            "schema": { "type": "integer", "default": 10 },
            "description": "Seconds to wait before killing the container"
          }
        ],
        "responses": {
          "200": {
            "description": "Container stopped",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Container" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/containers/{id}/restart": {
      "post": {
        "summary": "Restart a container",
        "tags": ["containers"],
        "operationId": "restartContainer",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" },
          {
            "name": "timeout",
            "in": "query",
            "schema": { "type": "integer", "default": 10 },
            "description": "Seconds to wait before killing the container"
          }
        ],
        "responses": {
          "200": {
            "description": "Container restarted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Container" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/containers/{id}/logs": {
      "get": {
        "summary": "Stream container logs as Server-Sent Events",
        "tags": ["containers"],
        "operationId": "streamContainerLogs",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" },
          {
            "name": "tail",
            "in": "query",
            "schema": { "type": "integer", "default": 100 },
            "description": "Number of lines to show from end of logs before streaming"
          },
          {
            "name": "follow",
            "in": "query",
            "schema": { "type": "boolean", "default": true },
            "description": "Whether to stream new log lines"
          },
          {
            "name": "timestamps",
            "in": "query",
            "schema": { "type": "boolean", "default": true }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE log stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream. Each event is a JSON-encoded LogEntry."
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/containers/{id}/stats": {
      "get": {
        "summary": "Stream container resource statistics as Server-Sent Events",
        "tags": ["containers"],
        "operationId": "streamContainerStats",
        "parameters": [
          { "$ref": "#/components/parameters/ContainerId" },
          {
            "name": "stream",
            "in": "query",
            "schema": { "type": "boolean", "default": true },
            "description": "If false, return a single snapshot and close"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stats stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream. Each event is a JSON-encoded ContainerStats."
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ContainerId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Docker container ID (short or full 64-char hash)"
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request or validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256-signed JWT issued by /auth/login. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "Container": {
        "type": "object",
        "description": "Represents a Docker container managed by DCMS",
        "properties": {
          "id": { "type": "string", "description": "Docker container ID (64-char SHA256)" },
          "short_id": { "type": "string", "description": "First 12 characters of container ID" },
          "name": { "type": "string", "example": "my-nginx" },
          "image": { "type": "string", "example": "nginx:1.25-alpine" },
          "image_id": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["created", "running", "paused", "restarting", "removing", "exited", "dead"]
          },
          "state": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "started_at": { "type": "string", "format": "date-time" },
          "finished_at": { "type": "string", "format": "date-time" },
          "host_id": { "type": "string", "description": "DCMS host record ID for the Docker daemon" },
          "host_name": { "type": "string" },
          "namespace": { "type": "string", "description": "Logical namespace (from dcms.namespace label)" },
          "ports": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "container_port": { "type": "integer" },
                "host_port": { "type": "integer" },
                "protocol": { "type": "string", "enum": ["tcp", "udp"] },
                "host_ip": { "type": "string" }
              }
            }
          },
          "env": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Environment variables (redacted values for secrets)"
          },
          "labels": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "networks": {
            "type": "array",
            "items": { "type": "string" }
          },
          "volumes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "source": { "type": "string" },
                "destination": { "type": "string" },
                "mode": { "type": "string" }
              }
            }
          },
          "restart_count": { "type": "integer" },
          "restart_policy": { "type": "string" },
          "cpu_shares": { "type": "integer" },
          "memory_limit_bytes": { "type": "integer" }
        },
        "required": ["id", "name", "image", "status", "host_id", "created_at"]
      },
      "ContainerCreateRequest": {
        "type": "object",
        "description": "Payload to create a new container",
        "required": ["image", "host_id"],
        "properties": {
          "image": { "type": "string", "example": "nginx:1.25-alpine" },
          "name": { "type": "string", "description": "Optional container name; auto-generated if omitted" },
          "host_id": { "type": "string", "description": "Target DCMS host ID" },
          "namespace": { "type": "string" },
          "cmd": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Override the image CMD"
          },
          "entrypoint": {
            "type": "array",
            "items": { "type": "string" }
          },
          "env": {
            "type": "array",
            "items": { "type": "string" },
            "description": "KEY=VALUE pairs"
          },
          "ports": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["container_port"],
              "properties": {
                "container_port": { "type": "integer" },
                "host_port": { "type": "integer" },
                "protocol": { "type": "string", "enum": ["tcp", "udp"], "default": "tcp" },
                "host_ip": { "type": "string", "default": "0.0.0.0" }
              }
            }
          },
          "volumes": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["source", "destination"],
              "properties": {
                "source": { "type": "string" },
                "destination": { "type": "string" },
                "mode": { "type": "string", "enum": ["ro", "rw"], "default": "rw" }
              }
            }
          },
          "networks": {
            "type": "array",
            "items": { "type": "string" }
          },
          "labels": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "restart_policy": {
            "type": "string",
            "enum": ["no", "always", "on-failure", "unless-stopped"],
            "default": "no"
          },
          "cpu_shares": { "type": "integer", "description": "Relative CPU weight (1-1024)" },
          "cpu_quota": { "type": "integer", "description": "CPU CFS quota in microseconds" },
          "memory_limit_bytes": { "type": "integer", "description": "Memory limit; 0 = unlimited" },
          "memory_swap_bytes": { "type": "integer" },
          "privileged": { "type": "boolean", "default": false },
          "read_only_rootfs": { "type": "boolean", "default": false },
          "user": { "type": "string", "description": "UID or UID:GID to run the container as" },
          "working_dir": { "type": "string" },
          "auto_start": { "type": "boolean", "default": true, "description": "Start container immediately after creation" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned for all 4xx and 5xx responses",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "description": "Machine-readable error code", "example": "CONTAINER_NOT_FOUND" },
          "message": { "type": "string", "description": "Human-readable error description" },
          "details": {
            "type": "array",
            "description": "Field-level validation errors",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "issue": { "type": "string" }
              }
            }
          },
          "trace_id": { "type": "string", "description": "Distributed trace ID for log correlation" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "description": "Pagination envelope — concrete responses extend this with a typed `data` array",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer", "description": "Total number of matching items" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_prev": { "type": "boolean" }
        },
        "required": ["page", "limit", "total", "total_pages"]
      }
    }
  }
}